+++ lib/errors/platform_exception/custom_platform_exception.dart
@@ -9,6 +9,7 @@ class CustomPlatformException extends BaseException {
     super.error,
     super.st,
     super.errorDetailsFromDependency,
+    super.process,
     required super.errorEnum,
   });
 }